requirements.txt
gitingest>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.0.0